src/requirements.txt
# Django Stuff
asgiref==3.2.3
Django==3.0.6
djangorestframework==3.10.3
django-cors-headers==3.2.1
pytz==2019.3
sqlparse==0.3.0

# SQL Conections
mysqlclient==1.4.6
PyMySQL==0.9.3

# Other Stuff
PyYAML==5.1.2
PyJWT==1.7.1
